CURB-4267 Migrated Page and WriteReview pages
