+++ libraries/engage/components/Logo/spec.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import Logo from './index';
 
+jest.mock('./connector', () => Component => Component);
+
 describe('<Logo />', () => {
   it('should render an image', () => {
     const wrapper = shallow(<Logo />);
