+++ themes/theme-gmd/components/AppBar/presets/DefaultBar/components/WideBar/WideBar.jsx
@@ -3,9 +3,8 @@ import PropTypes from 'prop-types';
 import { css } from 'glamor';
 import { themeConfig } from '@shopgate/pwa-common/helpers/config';
 import { AppBar } from '@shopgate/pwa-ui-material';
-import { ResponsiveContainer } from '@shopgate/engage/components';
+import { ResponsiveContainer, Logo } from '@shopgate/engage/components';
 import { GlobalLocationSwitcher, FulfillmentSlotSwitcher } from '@shopgate/engage/locations';
-import Logo from 'Components/Logo';
 import { MAX_DESKTOP_WIDTH, DESKTOP_MENU_BAR_HEIGHT } from '../../../../../../constants';
 import Search from './Search';
 import Cart from './Cart';
