+++ themes/theme-ios11/pages/Page/components/Content/index.jsx
@@ -10,7 +10,7 @@ import { PAGE_ID_INDEX } from '@shopgate/pwa-common/constants/PageIDs';
 import Widgets from '@shopgate/pwa-common/components/Widgets';
 import { AppBar } from '@shopgate/pwa-ui-ios';
 import { DefaultBar, BackBar } from 'Components/AppBar/presets';
-import Logo from 'Components/Logo';
+import { Logo } from '@shopgate/engage/components';
 import widgets from 'Extensions/widgets';
 import connect from './connector';
